Refactor to single optimized app.py and improve performance
Consolidate all optimizations and features into a single app.py file, removing app_optimized.py, config.py, performance_monitor.py, and quick_start.py. Integrate Redis caching, database connection pooling, batch processing, and pre-computed word tokens directly into app.py. Update README with new single-file usage, dependencies, and deployment instructions. Enhance requirements.txt for latest dependencies and remove obsolete files.
