+++ requirements.txt
@@ -1,5 +1,5 @@
 flask==3.0.0
-openai>=1.12.0
+openai==1.99.3
 python-dotenv==1.0.0
 spacy==3.7.2
 urllib3>=2.0.0
